pyproject.toml
[tool.poetry]
name = "AlekSIS-App-CSVImport"
version = "2.0a1"
packages = [
    { include = "aleksis" }
]
readme = "README.rst"

description = "AlekSIS (School Information System) — App for SchILD-NRW interaction"
authors = ["Dominik George <dominik.george@teckids.org>"]
license = "EUPL-1.2"
homepage = "https://aleksis.org/"
repository = "https://edugit.org/AlekSIS/Onboarding/AlekSIS-App-CSVImport"
documentation = "https://aleksis.edugit.io/AlekSIS/docs/html/"
classifiers = [
    "Environment :: Web Environment",
    "Intended Audience :: Education",
    "Topic :: Education"
]

[tool.poetry.dependencies]
python = "^3.7"
AlekSIS = "^2.0a2"
AlekSIS-App-Chronos = "^2.0a2"
pandas = "^1.0.0"
phonenumbers = "^8.10"
dateparser = "^0.7.6"

[tool.poetry.dev-dependencies]
sphinx = "^3.0"
sphinxcontrib-django = "^0.5.0"
sphinx-autodoc-typehints = "^1.7"
django-stubs = "^1.1"
pytest = "^5.3"
pytest-django = "^3.7"
pytest-django-testing-postgresql = "^0.1"
selenium = "^3.141.0"
safety = "^1.8.5"
flake8 = "^3.7.9"
flake8-django = "^1.0.0"
flake8-fixme = "^1.1.1"
flake8-mypy = "^17.8.0"
flake8-bandit = "^2.1.2"
flake8-builtins = "^1.4.1"
flake8-docstrings = "^1.5.0"
flake8-rst-docstrings = "^0.0.13"
black = "^19.10b0"
flake8-black = "^0.1.1"
isort = "^4.3.21"
flake8-isort = "^3.0.0"
pytest-cov = "^2.8.1"
pytest-sugar = "^0.9.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
